% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisync.R
\name{logisync}
\alias{logisync}
\title{\code{logisync}}
\usage{
logisync(seu_obj, csv, soup_k, output_col='FinalAssignment', res=FALSE)
}
\arguments{
\item{seu_obj}{The input Seurat object. Must contain hash assay named \code{HTO} or \code{hto} in dgCMatrix format.
Must contain Souporcell genotype assay(s) named \code{GENO} or \code{geno} with the desired k value (ex: \code{GENO5}).
Assumes Souporcell was run renaming all multiplet designations to 'multiplet' (see \code{\link[viewmastR:add_souporcell_seurat]{viewmastR::add_souporcell_seurat()}}
documentation).}

\item{csv}{The input hash-sample csv file path, or a hash-sample dataframe. The dataframe must contain \code{Hash} and \code{Sample} columns.}

\item{soup_k}{The desired Souporcell run, indicating the number of genotypes detected. The appropriate
number of genotypes expected for the data should be determined prior to running kmeansync.}

\item{output_col}{The name for the sample assignments column that will be added to the output Seurat object.
Default is \code{FinalAssignment}.}

\item{res}{Boolean indicating additional results output. Default is FALSE. If TRUE, Seurat object with sample
and hash labels will be returned, along with a dataframe with False Discovery Rate (FDR) and a statistical significance
indicator (* when FDR < 0.1). This table shows statistical evidence that the assigned hash has an effect on the genotype.
Also returns a dataframe linking genotypes to samples, and a graph of the average marginal effects for each genotype.}
}
\value{
seu_obj Returns Seurat object with sample and hash labels by default. Note that multiplets have been excluded.
If \code{res=TRUE}, additional dataframes and graphs are returned in a list along with the Seurat object.
}
\description{
\code{logisync} To demultiplex pooled samples, runs logistic regression on hash data from an input Seurat object.
Binary logistic regression is run for each genotype, and hashes are assigned based on which hash has the largest influence on
the genotype (via average marginal effect). Then, genotypes can be linked to samples via the input hash-sample csv.
Note that binary Firth logistic regression will be run for genotypes with small proportions to account for rare cases and reduce bias.
}
\examples{
\dontrun{

 output_list4 <- logisync(seu_ABCD, csv='/path/to/hash_sampleABCD.csv', soup_k=4, res=TRUE)
 
 output_list3 <- logisync(seu_ABC, csv=hash_ABC_df, soup_k=3)
 
 seu5 <- logisync(seu_DEF, csv='/path/to/hash_sampleDEF.csv', soup_k=5, 
                  output_col='Sample_Assignment') 
  
}

}
