% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansync.R
\name{kmeansync}
\alias{kmeansync}
\title{\code{kmeansync}}
\usage{
kmeansync(seu_obj, csv, soup_k, conf=0.8, output_col='FinalAssignment', res=FALSE)
}
\arguments{
\item{seu_obj}{The input Seurat object. Must contain hash assay named 'HTO' or 'hto' in dgCMatrix format.
Must contain Souporcell genotype assay(s) named GENO or geno with the desired k value (ex: 'GENO5').
Assumes Souporcell was run renaming all multiplet designations to 'multiplet' (see viewmastR
add_souporcell_seurat documentation for rename_assignments).}

\item{csv}{The input hash-sample csv file path. The csv must contain \code{Hash} and \code{Sample} columns.}

\item{soup_k}{The desired Souporcell run, indicating the number of genotypes detected. The appropriate
number of genotypes expected for the data should be determined prior to running kmeansync.}

\item{conf}{The confidence for apriori association rules, a measure of how often rules are found to be true.
Default is 0.8, as recommended.}

\item{output_col}{The name for the sample assignments column that will be added to the output Seurat object.
Default is 'FinalAssignment'.}

\item{res}{Boolean indicating additional results output. Default is FALSE. If TRUE, Seurat object with sample
and hash labels will be returned, along with a dataframe linking kmeans clusters to genotypes, a dataframe
linking genotypes to samples, the association rules, the silhouette graph with optimal kmeans clusters, and
a graph of kmeans clustering, shaded with average hash per cluster.}
}
\value{
seu_obj Returns Seurat object with sample and hash labels by default. Note that multiplets have been excluded.
If res=TRUE, additional dataframes and graphs are returned as well.
}
\description{
\code{kmeansync} To demultiplex pooled samples, runs kmeans on hash data from an input Seurat object.
The optimal number of clusters for kmeans is selected with the silhouette method. Hashes are assigned to kmeans clusters
based on the average hash enrichment in each cluster. Apriori association rules analysis links the hashes from kmeans
clusters to Souporcell genotypes. Then, genotypes can be linked to samples via the input hash-sample csv.
}
\examples{
\dontrun{

 output_list4 <- kmeansync(seu, csv='/path/to/hash_sampleABCD.csv', soup_k=4, res=TRUE)
 
 seu6 <- kmeansync(seu, csv='/path/to/hash_sampleDEF.csv', soup_k=6, output_col='Sample_Assignment') 
  
}

}
